jack version
