+++ metta/eval/analysis.py
@@ -7,7 +7,7 @@ from tabulate import tabulate
 from metta.eval.analysis_config import AnalysisConfig
 from metta.eval.eval_stats_db import EvalStatsDB
 from metta.rl.checkpoint_manager import CheckpointManager
-from mettagrid.util.file import local_copy
+from metta.utils.file import local_copy
 
 
 def analyze(policy_uri: str, config: AnalysisConfig) -> None:
