+++ metta/eval/dashboard_data.py
@@ -6,8 +6,8 @@ from typing import Dict, List
 from pydantic import BaseModel
 
 from metta.sim.simulation_stats_db import SimulationStatsDB
+from metta.utils.file import write_data
 from mettagrid.config import Config
-from mettagrid.util.file import write_data
 
 logger = logging.getLogger(__name__)
