+++ metta/map/terrain_from_numpy.py
@@ -10,8 +10,8 @@ from filelock import FileLock
 from pydantic import ConfigDict, Field
 
 from metta.common.util.log_config import getRankAwareLogger
+from metta.utils.uri import ParsedURI
 from mettagrid.map_builder.map_builder import GameMap, MapBuilder, MapBuilderConfig
-from mettagrid.util.uri import ParsedURI
 
 logger = getRankAwareLogger(__name__)
