+++ metta/rl/training/torch_profiler.py
@@ -14,7 +14,7 @@ import wandb
 from metta.common.wandb.context import WandbRun
 from metta.rl.training import ComponentContext, TrainerComponent
 from metta.rl.utils import should_run
-from mettagrid.util.file import http_url, is_public_uri, write_file
+from metta.utils.file import http_url, is_public_uri, write_file
 
 logger = logging.getLogger(__name__)
