+++ metta/sim/simulation.py
@@ -21,13 +21,13 @@ from metta.common.util.heartbeat import record_heartbeat
 from metta.rl.checkpoint_manager import CheckpointManager
 from metta.rl.training.training_environment import EnvironmentMetaData
 from metta.rl.vecenv import make_vecenv
+from metta.sim.replay_writer import S3ReplayWriter
 from metta.sim.simulation_config import SimulationConfig
 from metta.sim.simulation_stats_db import SimulationStatsDB
 from metta.sim.stats import DuckDBStatsWriter
 from metta.sim.thumbnail_automation import maybe_generate_and_upload_thumbnail
 from metta.sim.utils import get_or_create_policy_ids
 from mettagrid import MettaGridEnv, dtype_actions
-from mettagrid.util.replay_writer import ReplayWriter
 
 SYNTHETIC_EVAL_SUITE = "training"
 
@@ -67,7 +67,7 @@ class Simulation:
         sim_stats_dir.mkdir(parents=True, exist_ok=True)
         self._stats_dir = sim_stats_dir
         self._stats_writer = DuckDBStatsWriter(sim_stats_dir)
-        self._replay_writer = ReplayWriter(replay_dir)
+        self._replay_writer = S3ReplayWriter(replay_dir)
         self._device = device
 
         self._display_name = f"{cfg.suite}/{cfg.name}"
