+++ metta/utils/__init__.py
@@ -1,1 +1,5 @@
 """Utility modules for metta."""
+
+from metta.utils import file, uri
+
+__all__ = ["file", "uri"]
