+++ packages/cogames/pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "cogames"
-version = "0.3.2"
+version = "0.3.5"
 description = "Multi-agent cooperative games"
 readme = "README.md"
 requires-python = ">=3.11.7"
