+++ packages/mettagrid/pyproject.toml
@@ -11,7 +11,8 @@ backend-path = ["."]
 
 [project]
 name = "mettagrid" # Do not change! Has to be 'mettagrid' for our PyPi package
-dynamic = ["version"]
+# dynamic = ["version"]
+version = "0.2.0.11"
 description = "A fast grid-based open-ended MARL environment"
 authors = [{ name = "David Bloomin", email = "daveey@gmail.com" }]
 requires-python = ">=3.11,<3.12"
